kafka_consumer: extract shared cache helpers into EventCacheMixin
Move _get_items_to_fetch, _mark_items_fetched, _get_events_to_send,
_get_tags, and _original_cluster_id_field into a new EventCacheMixin so
ClusterMetadataCollector and KafkaConnectCollector share one copy.

Also removes the dashboard JSON (shipped too early, no product sign-off)
and renames CONFIGS_REFRESH_INTERVAL/JITTER to private _configs_refresh_*
to match the convention already used in KafkaConnectCollector.
