+++ kafka_consumer/datadog_checks/kafka_consumer/connectors.py
@@ -2,16 +2,16 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 
-import hashlib
 import json
 import logging
-import random
 import re
 import time
 from typing import TYPE_CHECKING, Any
 from urllib.parse import quote
 
 import requests
+
+from datadog_checks.kafka_consumer.event_cache_mixin import EventCacheMixin
 
 if TYPE_CHECKING:
     from datadog_checks.kafka_consumer.config import KafkaConfig
@@ -22,7 +22,6 @@ CONNECTOR_CONFIG_CACHE_KEY = 'kafka_connector_config_cache'
 CONNECTOR_PLUGINS_CACHE_KEY = 'kafka_connector_plugins_cache'
 CONNECTOR_PLUGINS_EVENT_CACHE_KEY = 'kafka_connector_plugins_event_cache'
 CONNECTOR_CONFIG_CACHE_MAX_SIZE = 5_000
-EVENT_CACHE_TTL = 3600
 
 IMPORTANT_CONNECTOR_CONFIG_KEYS = [
     'connector.class',
@@ -93,7 +92,7 @@ def _fetch_oidc_token(oauth_config: dict, session: requests.Session, timeout: fl
     return access_token, expires_at
 
 
-class KafkaConnectCollector:
+class KafkaConnectCollector(EventCacheMixin):
     """Collects Kafka Connect connector metrics and config events."""
 
     def __init__(self, check, config: 'KafkaConfig', log: logging.Logger) -> None:
@@ -313,7 +312,9 @@ class KafkaConnectCollector:
 
         safe_url = quote(url, safe='')
         cache_key = f'{CONNECTOR_CONFIG_CACHE_KEY}:{safe_url}'
-        connectors_to_emit = self._get_events_to_send(cache_key, connector_contents)
+        connectors_to_emit = self._get_events_to_send(
+            cache_key, connector_contents, max_cache_size=CONNECTOR_CONFIG_CACHE_MAX_SIZE
+        )
 
         collection_timestamp = int(time.time() * 1000)
         for name in connectors_to_emit:
@@ -371,7 +372,9 @@ class KafkaConnectCollector:
             'plugins': plugins,
         }
         content = json.dumps(content_dict, sort_keys=True)
-        if self._get_events_to_send(event_cache_key, {'plugins': content}):
+        if self._get_events_to_send(
+            event_cache_key, {'plugins': content}, max_cache_size=CONNECTOR_CONFIG_CACHE_MAX_SIZE
+        ):
             event = json.loads(content)
             event['collection_timestamp'] = int(time.time() * 1000)
             self.check.event_platform_event(json.dumps(event), 'data-streams-message')
@@ -383,114 +386,3 @@ class KafkaConnectCollector:
         url = f'{base}/connect/v1/environments/{env_id}/clusters/{cc_cluster_id}'
         self._collect_rest(url, cluster_id)
 
-    def _get_items_to_fetch(self, cache_key: str, item_keys: list[str]) -> list[str]:
-        current_time = time.time()
-        items_to_fetch = []
-
-        try:
-            cached_str = self.check.read_persistent_cache(cache_key)
-            cache_dict = json.loads(cached_str) if cached_str else {}
-        except Exception as e:
-            self.log.debug("Could not read cache %s: %s", cache_key, e)
-            cache_dict = {}
-
-        for item_key in item_keys:
-            expire_at = cache_dict.get(item_key, 0)
-            if current_time >= expire_at:
-                items_to_fetch.append((expire_at, item_key))
-
-        items_to_fetch.sort()
-        return [item_key for _, item_key in items_to_fetch]
-
-    def _mark_items_fetched(
-        self,
-        cache_key: str,
-        item_keys: list[str],
-        ttl_base: float | None = None,
-        ttl_jitter: float | None = None,
-        max_cache_size: int | None = None,
-    ) -> None:
-        if ttl_base is None:
-            ttl_base = self._configs_refresh_interval
-        if ttl_jitter is None:
-            ttl_jitter = self._configs_refresh_jitter
-
-        current_time = time.time()
-
-        try:
-            cached_str = self.check.read_persistent_cache(cache_key)
-            cache_dict = json.loads(cached_str) if cached_str else {}
-        except Exception as e:
-            self.log.debug("Could not read cache %s for update: %s", cache_key, e)
-            cache_dict = {}
-
-        for item_key in item_keys:
-            ttl = ttl_base + random.uniform(0, ttl_jitter)
-            cache_dict[item_key] = current_time + ttl
-
-        if max_cache_size and len(cache_dict) > max_cache_size:
-            sorted_keys = sorted(cache_dict, key=lambda k: cache_dict[k])
-            for key in sorted_keys[: len(cache_dict) - max_cache_size]:
-                del cache_dict[key]
-
-        try:
-            self.check.write_persistent_cache(cache_key, json.dumps(cache_dict))
-        except Exception as e:
-            self.log.debug("Could not write cache %s: %s", cache_key, e)
-
-    def _get_events_to_send(
-        self, cache_key: str, items: dict[str, str], max_cache_size: int | None = None
-    ) -> list[str]:
-        if not items:
-            return []
-
-        current_time = time.time()
-        events_to_send = []
-
-        try:
-            cached_str = self.check.read_persistent_cache(cache_key)
-            cache_dict = json.loads(cached_str) if cached_str else {}
-        except Exception as e:
-            self.log.debug("Could not read cache %s: %s", cache_key, e)
-            cache_dict = {}
-
-        for item_key, event_content in items.items():
-            current_hash = hashlib.sha256(event_content.encode('utf-8')).hexdigest()
-            cached_entry = cache_dict.get(item_key)
-
-            if (
-                not cached_entry
-                or cached_entry.get('hash', '') != current_hash
-                or current_time >= cached_entry.get('expire_at', 0)
-            ):
-                events_to_send.append(item_key)
-                cache_dict[item_key] = {
-                    'hash': current_hash,
-                    'expire_at': current_time + EVENT_CACHE_TTL,
-                }
-
-        if events_to_send:
-            ceiling = max_cache_size if max_cache_size is not None else CONNECTOR_CONFIG_CACHE_MAX_SIZE
-            if len(cache_dict) > ceiling:
-                sorted_keys = sorted(cache_dict, key=lambda k: cache_dict[k].get('expire_at', 0))
-                for key in sorted_keys[: len(cache_dict) - ceiling]:
-                    del cache_dict[key]
-            try:
-                self.check.write_persistent_cache(cache_key, json.dumps(cache_dict))
-            except Exception as e:
-                self.log.debug("Could not write cache %s: %s", cache_key, e)
-
-        return events_to_send
-
-    def _get_tags(self, cluster_id: str | None = None) -> list[str]:
-        tags = list(self.config._custom_tags)
-        if cluster_id:
-            tags.append(f'kafka_cluster_id:{cluster_id}')
-            if self.config._kafka_cluster_id_override:
-                tags.append(f'original_kafka_cluster_id:{self.config._auto_detected_cluster_id}')
-        return tags
-
-    def _original_cluster_id_field(self) -> dict[str, str]:
-        if self.config._kafka_cluster_id_override:
-            return {'original_kafka_cluster_id': self.config._auto_detected_cluster_id}
-        return {}
